Make it all in one file for development to be easier to avoid circular imports >:(, fix #20 and #17
