--- EpikCord/colours.py
@@ -1,144 +1,0 @@
-"""
-Some parts of the code is done by discord.py and their amazing team of contributers
-The MIT License (MIT)
-Copyright © 2015-2021 discord.py
-Copyright © 2021-present Epikhost
-Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the “Software”), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE."""
-
-
-
-
-
-import random
-import colorsys
-from typing import (Any, Optional, Tuple,Type ,TypeVar, Union)
-
-__all__ = ('Color', 
-'Colour',
-
-) #include british and american english
-
-CT = TypeVar('CT', bound='Colour')
-class Colour:
-    #some of this code is sourced from discord.py, rest assured all the colors are different from discord.py
-    __slots__ = ('value',)
-
-    def __init__(self, value: int):
-        if not isinstance(value, int):
-            raise TypeError(f'Expected int parameter, received {value.__class__.__name__} instead.')
-
-        self.value: int = value
-
-    def _get_byte(self, byte: int) -> int:
-        return (self.value >> (8 * byte)) & 0xff
-
-    def __eq__(self, other: Any) -> bool:
-        return isinstance(other, Colour) and self.value == other.value
-
-    def __ne__(self, other: Any) -> bool:
-        return not self.__eq__(other)
-
-    def __str__(self) -> str:
-        return f'#{self.value:0>6x}'
-
-    def __int__(self) -> int:
-        return self.value
-
-    def __repr__(self) -> str:
-        return f'<Colour value={self.value}>'
-
-    def __hash__(self) -> int:
-        return hash(self.value)
-
-    @property
-    def r(self) ->int:
-        """Return the red component in rgb"""
-        return self._get_byte(2)
-
-    @property
-    def g(self) ->int:
-        """Return the green component in rgb"""
-        return self._get_byte(1)
-
-    @property
-    def b(self) ->int:
-        """Return the blue component in rgb"""
-        return self._get_byte(0)
-
-    def to_rgb(self) -> Tuple[int, int, int]:
-        """returns an rgb color as a tuple"""
-        return (self.r, self.g,self.b)
-
-    @classmethod
-    def from_rgb(cls: Type[CT], r: int, g: int, b: int) -> CT:
-        """Constructs a :class:`Colour` from an RGB tuple."""
-        return cls((r << 16) + (g << 8) + b)
-
-    @classmethod
-    def lime_green(cls: Type[CT]) -> CT:
-        """Returns a color of lime green"""
-        return cls(0x00ff01)
-
-    @classmethod
-    def light_green(cls: Type[CT]) -> CT:
-        """Returns a color of light green"""
-        return cls(0x00ff22)
-
-    @classmethod
-    def dark_green(cls: Type[CT]) -> CT:
-        """Returns a color of dark green"""
-        return cls(0x00570a)
-    
-    @classmethod
-    def light_blue(cls: Type[CT]) -> CT:
-        """Returns a color of light blue"""
-        return cls(0x00ff01)
-
-    @classmethod
-    def dark_blue(cls: Type[CT]) -> CT:
-        """Returns a color of dark blue"""
-        return cls(0x0a134b)
-
-    @classmethod
-    def light_red(cls: Type[CT]) -> CT:
-        """Returns a color of light red"""
-        return cls(0xaa5b54)
-
-    @classmethod
-    def dark_red(cls: Type[CT]) -> CT:
-        """Returns a color of dark red"""
-        return cls(0x4c0000)
-    
-    @classmethod
-    def black(cls: Type[CT]) -> CT:
-        """Returns a color of black"""
-        return cls(0x00000)
-
-    default=black
-    @classmethod
-    def white(cls: Type[CT]) -> CT:
-        return cls(0xffffff)
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-Color=Colour
-
-    
-    
-
-
