--- EpikCord/section.py
@@ -1,26 +1,0 @@
-from .slash_command import Subcommand, SubCommandGroup, StringOption, IntegerOption, BooleanOption, UserOption, ChannelOption, RoleOption, MentionableOption, NumberOption
-from typing import (
-    Optional,
-    List,
-    Union
-)
-
-class Section:
-    def __init__(self):
-        self.commands = {}
-        self.events = {}
-    
-    def event(self, event_name: str):
-        def register_event(func):
-            self.events[event_name] = func
-        return register_event
-
-    def slash_command(self,*, name: str, description: Optional[str], options: List[Union[Subcommand, SubCommandGroup, StringOption, IntegerOption, BooleanOption, UserOption, ChannelOption, RoleOption, MentionableOption, NumberOption]]):
-        def register_slash_command(func):
-            self.commands[name] = {
-                "callback": func,
-                "name": name,
-                "description": description,
-                "options": options
-            }
-        return register_slash_command
