--- EpikCord/thread.py
@@ -1,68 +1,0 @@
-from typing import (
-    List,
-    Optional
-)
-from .exceptions import ThreadArchived, NotFound404
-
-class ThreadMember:
-    def __init__(self, data: dict):
-        self.id: str = data["user_id"]
-        self.thread_id: str = data["thread_id"]
-        self.join_timestamp: str = data["join_timestamp"]
-        self.flags: int = data["flags"]
-
-class Thread:
-    def __init__(self, client, data: dict):
-        super().__init__(client, data)
-        self.owner_id: str = data["owner_id"]
-        self.message_count: int = data["message_count"]
-        self.member_count: int = data["member_count"]
-        self.archived: bool = data["archived"]
-        self.auto_archive_duration: int = data["auto_archive_duration"]
-        self.archive_timestamp: str = data["archive_timestamp"]
-        self.locked: bool = data["locked"]
-    
-    async def join(self):
-        if self.archived:
-            raise ThreadArchived("This thread has been archived so it is no longer joinable")
-        response = await self.client.http.put(f"/channels/{self.id}/thread-members/@me")
-        return await response.json()
-    
-    async def add_member(self, member_id: str):
-        if self.archived:
-            raise ThreadArchived("This thread has been archived so it is no longer joinable")
-        
-        response = await self.client.http.put(f"/channels/{self.id}/thread-members/{member_id}")
-        return await response.json()
-    
-    async def leave(self):
-        if self.archived:
-            raise ThreadArchived("This thread has been archived so it is no longer leaveable")
-        response = await self.client.http.delete(f"/channels/{self.id}/thread-members/@me")
-        return await response.json()
-    
-    async def remove_member(self, member_id: str):
-        if self.archived:
-            raise ThreadArchived("This thread has been archived so it is no longer leaveable")
-        
-        response = await self.client.http.delete(f"/channels/{self.id}/thread-members/{member_id}")
-        return await response.json()
-    
-    async def fetch_member(self, member_id: str) -> ThreadMember:
-        response = await self.client.http.get(f"/channels/{self.id}/thread-members/{member_id}")
-        if response.status == 404:
-            raise NotFound404("The member you are trying to fetch does not exist")
-        return ThreadMember(await response.json())
-    
-    async def list_members(self) -> List[ThreadMember]:
-        response = await self.client.http.get(f"/channels/{self.id}/thread-members")
-        return [ThreadMember(member) for member in await response.json()]
-
-    async def bulk_delete(self, message_ids: List[str], reason: Optional[str]) -> None:
-
-        if reason:
-            headers = self.client.http.headers.copy()
-            headers["X-Audit-Log-Reason"] = reason
-            
-        response = await self.client.http.post(f"channels/{self.id}/messages/bulk-delete", data={"messages": message_ids}, headers=headers)
-        return await response.json()
